+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-* Fix [AppleSkin](https://modrinth.com/mod/appleskin) not showing food nutrition in tooltips
-* Fix orth coin worth sometimes showing up with a white shadow
-* Mod compatibility mixins no longer try to load if the mod is not installed
+* Add config option:
+  * `Hide Gear Ability Fails in Action Bar`